set random integer functions return values to 16 bit

+++ xlive/Blam/Engine/math/random_math.cpp
@@ -64,13 +64,13 @@ real32 _real_random_range(uint32* seed, real32 lower_bound, real32 upper_bound)
 	return REAL_RANDOM_RANGE(*seed, lower_bound, upper_bound - lower_bound);
 }
 
-int32 _random_range(uint32* seed, int16 lower_bound, int16 upper_bound)
+int16 _random_range(uint32* seed, int16 lower_bound, int16 upper_bound)
 {
 	*seed = RANDOM_NEW_SEED(*seed);
 	return INTEGER_RANDOM_RANGE(*seed, lower_bound, upper_bound - lower_bound);
 }
 
-int32 _random_integer(uint32* seed)
+int16 _random_integer(uint32* seed)
 {
 	*seed = RANDOM_NEW_SEED(*seed);
 	return INTEGER_RANDOM(*seed);
